+++ internal/storage/tribonacciStorage.go
@@ -1,6 +1,6 @@
 package storage
 
 type TribonacciStorage interface {
-	Get(signature []float32, n int64) ([]float32, error) // slice, err
+	Get(signature []float32, n int) ([]float32, error)
 	Set(signature []float32, values []float32) error
 }
